% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{ApplyClusteringResultsToSeurat}
\alias{ApplyClusteringResultsToSeurat}
\title{ApplyClusteringResultsToSeurat}
\usage{
ApplyClusteringResultsToSeurat(sourceSeuratObj, targetSeuratObj)
}
\arguments{
\item{sourceSeuratObj}{The source seurat object, containing clustering results}

\item{targetSeuratObj}{The target seurat object, where clustering results will be joined}
}
\description{
This takes a seuratObject with TCR clustering results calculated, and then joins the TCR family IDs from this object to the target object, where this join is performed using the CDR3, V and J data.
}
