% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateTcrDistances.R
\name{CalculateTcrDistances}
\alias{CalculateTcrDistances}
\title{CalculateTcrDistances}
\usage{
CalculateTcrDistances(
  inputData = NULL,
  organism = "human",
  chains = c("TRA", "TRB"),
  spikeInDataframe = NULL,
  minimumCloneSize = 2,
  calculateChainPairs = FALSE,
  debugMode = FALSE,
  verbose = FALSE,
  method = "tcrdist3"
)
}
\arguments{
\item{inputData}{Either a seurat Object containing TCR information or a dataframe containing metadata}

\item{organism}{Organism to use for reference segments. Default is 'human'.}

\item{chains}{Vector of TCR chains to include in the analysis. Default is c("TRA", "TRB").}

\item{spikeInDataframe}{An optional data frame containing additional CDR3s and gene segments to be included in the clustering, such as published clonotypes.}

\item{minimumCloneSize}{Minimum number of clones per subject to include in the analysis. Default is 2.}

\item{calculateChainPairs}{Boolean controlling whether to compute joint multi-chain distance matrices for observed chain combinations.}

\item{debugMode}{Boolean that will result in more debugging information to be printed}

\item{verbose}{Boolean controlling whether to print additional information about procesing}

\item{method}{The method for clustering. Currently only tcrdist3 is supported}
}
\description{
This function calculates pairwise TCR distances and stores them in a Seurat Object.
}
\examples{
\dontrun{
  resultList <- CalculateTcrDistances(inputData = seuratObj@meta.data,
              chains = c("TRA", "TRB"),
              minimumCloneSize = 2,
              calculateChainPairs = FALSE,
              verbose = FALSE)

  # Example with calculateChainPairs analysis
  resultList <- CalculateTcrDistances(inputData = seuratObj,
              chains = c("TRA", "TRB", "TRG", "TRD"),
              calculateChainPairs = TRUE,
              verbose = TRUE)

    spikeInDataframe <- data.frame(CloneNames = rep(1:3),
                                 TRA_V = c("TRAV1-2", "TRAV1-2", "TRAV1-2"),
                                 TRA_J = c("TRAJ33", "TRAJ20", "TRAJ33"),
                                 TRA = c("CAVRDSNYQLIW", "CAVSLQDYKLSF", "CAVRDSNYQLIW"),
                                 TRB_V = c("TRBV6-4", "TRBV6-4", "TRBV6-4"),
                                 TRB_J = c("TRBJ1-1", "TRBJ2-1", "TRBJ2-3"),
                                 TRB = c("CASSAAAAAAAAFF", "CASSVVVVVVVVQF", "CASSWWWWWWWWQY"))
}

}
