% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{.DianaClustering}
\alias{.DianaClustering}
\title{.DianaClustering}
\usage{
.DianaClustering(distanceMatrix, cutHeight = 20, verbose = FALSE)
}
\arguments{
\item{distanceMatrix}{Distance matrix (as matrix or dist object)}

\item{cutHeight}{Height at which to cut the dendrogram. Default is 20.}

\item{verbose}{Boolean indicating whether to print verbose output. Default is FALSE.}
}
\value{
A list containing the clustering assignments and dendrogram object
}
\description{
Internal function to perform DIANA (Divisive Analysis) hierarchical clustering
on a distance matrix and cut the dendrogram at a specified height.
}
\keyword{internal}
