% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatMetadata.R
\name{.FormatMetadata}
\alias{.FormatMetadata}
\title{Format metadata for processing}
\usage{
.FormatMetadata(
  metadata,
  chains = c("TRA", "TRB"),
  organism = "human",
  calculateChainPairs = TRUE,
  spikeInDataframe = NULL,
  verbose = FALSE
)
}
\arguments{
\item{metadata}{Data frame containing metadata.}

\item{chains}{TCR chains to include in the analysis. TRA/TRB supported and tested, but others likely work.}

\item{organism}{Organism to use for tcrDist3. Default is 'human'.}

\item{calculateChainPairs}{If true, this will prepare the columns needed for A/B and/or  G/D (depending on values of chains}

\item{spikeInDataframe}{Data frame containing spike-in data. Default is NULL. See examples for formatting requirements.}

\item{verbose}{Boolean controlling whether to display processing steps. Default is FALSE.}
}
\value{
a properly formatted metadata dataframe.
}
\description{
This function formats a seurat object's metadata (with TCR information appended) for tcrDist3 distance caluclations.
}
\examples{
\dontrun{
spikeInDataframe <- data.frame(CloneNames = rep(1:3),
                                 TRA_V = c("TRAV1-2", "TRAV1-2", "TRAV1-2"),
                                 TRA_J = c("TRAJ33", "TRAJ20", "TRAJ33"),
                                 TRA = c("CAVRDSNYQLIW", "CAVSLQDYKLSF", "CAVRDSNYQLIW"),
                                 TRB_V = c("TRBV6-4", "TRBV6-4", "TRBV6-4"),
                                 TRB_J = c("TRBJ1-1", "TRBJ2-1", "TRBJ2-3"),
                                 TRB = c("CASSAAAAAAAAFF", "CASSVVVVVVVVQF", "CASSWWWWWWWWQY"))
}
}
