% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{.ThresholdClustersBySize}
\alias{.ThresholdClustersBySize}
\title{.ThresholdClustersBySize}
\usage{
.ThresholdClustersBySize(
  clusterAssignments,
  minClusterSize = 2,
  verbose = FALSE
)
}
\arguments{
\item{clusterAssignments}{Named vector of cluster assignments (names are cell/clone IDs)}

\item{minClusterSize}{Minimum number of unique clones required to keep a cluster. Default is 2.}

\item{verbose}{Boolean indicating whether to print verbose output. Default is FALSE.}
}
\value{
Named vector of cluster assignments with singletons/small clusters removed (set to 0)
}
\description{
Internal function to filter clusters by minimum number of unique clones
}
\keyword{internal}
