% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{ExpandDistancesToMatchSeuratObj}
\alias{ExpandDistancesToMatchSeuratObj}
\title{ExpandDistancesToMatchSeuratObj}
\usage{
ExpandDistancesToMatchSeuratObj(seuratObj, chains)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{chains}{A single chain or vector of the joint chains to use}
}
\description{
This takes the distance matrix, which is one row/col per clone, and joined this with the seuratObj in order to make a larger matrix with one column/cell
}
