% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{GetDistanceMatrix}
\alias{GetDistanceMatrix}
\title{GetDistanceMatrix}
\usage{
GetDistanceMatrix(seuratObj_TCR, chains, cdr3Only = FALSE)
}
\arguments{
\item{seuratObj_TCR}{The seurat object, created by CalculateTcrDistances}

\item{chains}{The chain or chains to return (e.g., TRA or c('TRG', 'TRD'))}

\item{cdr3Only}{If true, it will return the distances calculated on CDR3 alone. The default uses distances based on CDR3+V/J}
}
\description{
Can be used to set the python executable used by tcrClustR
}
