% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletDistances.R
\name{PlotMixingProportions}
\alias{PlotMixingProportions}
\title{Plot Mixing Proportions per Cluster}
\usage{
PlotMixingProportions(dpResult, verbose = FALSE)
}
\arguments{
\item{dpResult}{A \code{tcrDirichletResult} object from
\code{\link{DirichletClusterAnalysis}}.}

\item{verbose}{Logical. Default \code{FALSE}.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Companion to \code{\link{PlotClusterMeans}}. Shows the
proportion of data points assigned to each DirichletProcess cluster as a dodged bar
chart, one bar per group. Clusters with high mixing weight and low mean
distance identify well-supported clonotype families.
}
\examples{
\dontrun{
dp <- DirichletClusterAnalysis(seuratObj, "TRA_fl", "cDNA_ID")
PlotMixingProportions(dp)
}

}
