% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{SetPythonExecutable}
\alias{SetPythonExecutable}
\title{SetPythonExecutable}
\usage{
SetPythonExecutable(pythonExecutable = "")
}
\arguments{
\item{pythonExecutable}{The path to the python executable}
}
\description{
Can be used to set the python executable used by tcrClustR
}
