% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{.DistanceMatrixToClusteredGraphs}
\alias{.DistanceMatrixToClusteredGraphs}
\title{.DistanceMatrixToClusteredGraphs}
\usage{
.DistanceMatrixToClusteredGraphs(
  seuratObj_TCR = NULL,
  kpcaComponents = 50,
  kpcaKernel = "rbfdot",
  partitionType = "CPMVertexPartition",
  proportionOfGraphAsNeighbors = 0.1,
  jaccardIndexThreshold = 0.1,
  resolutions = c(0.1, 0.2, 0.3),
  computeMultiChain = T
)
}
\value{
Single Chain and multi-chain Seurat objects
}
\description{
.DistanceMatrixToClusteredGraphs
}
