% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congaUtils.R
\name{FormatMetadataForConga}
\alias{FormatMetadataForConga}
\title{FormatMetadataForConga}
\usage{
FormatMetadataForConga(
  metadata,
  outputCsv = "./congaInput.csv",
  organism = "human",
  chains = NULL,
  cleanMetadata = T,
  minimumClonesPerSubject = 100,
  writeUnannotatedGeneSegmentsToFile = T,
  spikeInDataframe = NULL
)
}
\arguments{
\item{metadata}{Data frame containing metadata.}

\item{outputCsv}{Path to the output CSV file.}

\item{organism}{Organism to use for conga's TCR db. Default is 'human'.}

\item{chains}{TCR chains to include in the analysis. TRA/TRB supported and tested, but others likely work}

\item{cleanMetadata}{Boolean controlling whether to clean the metadata by removing rows with NA values or commas in the specified chains.}

\item{minimumClonesPerSubject}{Minimum number of clones per subject to include in the analysis. Default is 2.}

\item{writeUnannotatedGeneSegmentsToFile}{Boolean controlling whether to write unannotated gene segments to a file (filtered_(chain)_gene_segments.csv).}

\item{spikeInDataframe}{Data frame containing spike-in data. Default is NULL. See examples for formatting requirements.}
}
\description{
This function formats metadata for the conga pipeline.
}
\examples{
\dontrun{
spikeInDataframe <- data.frame(CloneNames = rep(1:3),
                                 TRA_V = c("TRAV1-2", "TRAV1-2", "TRAV1-2"),
                                 TRA_J = c("TRAJ33", "TRAJ20", "TRAJ33"),
                                 TRA = c("CAVRDSNYQLIW", "CAVSLQDYKLSF", "CAVRDSNYQLIW"),
                                 TRB_V = c("TRBV6-4", "TRBV6-4", "TRBV6-4"),
                                 TRB_J = c("TRBJ1-1", "TRBJ2-1", "TRBJ2-3"),
                                 TRB = c("CASSAAAAAAAAFF", "CASSVVVVVVVVQF", "CASSWWWWWWWWQY")
}
}
