% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrDist3Utils.R
\name{FormatMetadataForTcrDist3}
\alias{FormatMetadataForTcrDist3}
\title{Format metadata for tcrDist3}
\usage{
FormatMetadataForTcrDist3(
  metadata,
  outputCsv = "./tcrDist3Input.csv",
  chains = c("TRA", "TRB"),
  cleanMetadata = T,
  summarizeClones = T,
  imputeCloneNames = T,
  minimumClonesPerSubject = 100,
  writeUnannotatedGeneSegmentsToFile = T
)
}
\arguments{
\item{metadata}{Data frame containing metadata.}

\item{outputCsv}{Path to the output CSV file.}

\item{chains}{TCR chains to include in the analysis. TRA/TRB supported and tested, but others likely work}

\item{cleanMetadata}{Boolean controlling whether to clean the metadata by removing rows with NA values or commas in the specified chains.}

\item{summarizeClones}{Boolean controlling whether to summarize clones by SubjectId, TRA, TRB, TRA_V, TRA_J, TRB_V, and TRB_J.}

\item{imputeCloneNames}{Boolean controlling whether to impute clone names if they are missing.}

\item{minimumClonesPerSubject}{Minimum number of clones per subject to include in the analysis. Default is 2.}

\item{writeUnannotatedGeneSegmentsToFile}{Boolean controlling whether to write unannotated gene segments to a file (filtered_(chain)_gene_segments.csv).}
}
\description{
This function formats a seurat object's metadata (with TCR information appended) for tcrDist3 distance caluclations.
}
