% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformClusteringAndStoreResults.R
\name{PerformClusteringAndStoreResults}
\alias{PerformClusteringAndStoreResults}
\title{PerformClusteringAndStoreResults}
\usage{
PerformClusteringAndStoreResults(
  seuratObj,
  outputDir,
  outputPrefix = "TCR.clustering",
  chains,
  verbose = TRUE,
  minimumCloneSize
)
}
\arguments{
\item{seuratObj}{The Seurat object}

\item{outputDir}{Directory to write parquet files. Default = "./tcrclustering_output/".}

\item{outputPrefix}{Optional prefix for output filenames. Default = "clustering".}

\item{verbose}{Logical. Print progress. Default = TRUE.}
}
\value{
A seurat object with clustering information stored as assays
}
\description{
A wrapper method designed to provide a start-to-finish pipeline to cluster/store TCR families for a given seurat object
}
