% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTcrdist3.R
\name{RunTcrdist3}
\alias{RunTcrdist3}
\title{RunTcrdist3}
\usage{
RunTcrdist3(
  seuratObj,
  organism = "human",
  chains = c("TRA", "TRB"),
  minimumCloneSize = 2,
  calculateChainPairs = FALSE,
  debugTcrdist3 = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{seuratObj}{The seuratObj containing formatted metadata}

\item{organism}{Organism to use for tcrdist3. Default is 'human'.}

\item{chains}{Vector of TCR chains to include in the analysis. Default is c("TRA", "TRB").}

\item{minimumCloneSize}{Minimum number of clones per subject to include in the analysis. Default is 2.}

\item{calculateChainPairs}{Boolean controlling whether to compute joint multi-chain distance matrices for observed chain combinations.}

\item{debugTcrdist3}{Boolean controlling whether to run tcrdist3 in debug mode.}

\item{verbose}{Boolean controlling whether to display processing steps.}
}
\description{
This function runs the tcrdist3 pipeline on a set of CDR3 sequences in a Seurat Object.
}
