% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrDist3Wrapper.R
\name{RunTcrdist3}
\alias{RunTcrdist3}
\title{RunTcrdist3}
\usage{
RunTcrdist3(
  inputData = NULL,
  organism = "human",
  chains = c("TRA", "TRB"),
  spikeInDataframe = NULL,
  summarizeClones = T,
  imputeCloneNames = T,
  minimumClonesPerSubject = 2,
  multichain = FALSE,
  rdsOutputPath = "./tcrdist3DistanceMatrices/",
  pythonExecutable = NULL,
  debugTcrdist3 = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{inputData}{Either a seurat Object containing TCR information or a dataframe containing metadata}

\item{organism}{Organism to use for tcrdist3. Default is 'human'.}

\item{chains}{Vector of TCR chains to include in the analysis. Default is c("TRA", "TRB").}

\item{spikeInDataframe}{Data frame containing known CDR3s and gene segments to be included in the clustering. Default is NULL.}

\item{summarizeClones}{Pass-through boolean controlling whether to summarize clones into a frequency (by SubjectId, TRA, TRB, TRA_V, and TRB_J). Default is TRUE.}

\item{imputeCloneNames}{Pass-through boolean controlling whether to impute clone names if they are missing. Existing clone names will be inherited. Default is TRUE.}

\item{minimumClonesPerSubject}{Minimum number of clones per subject to include in the analysis. Default is 2.}

\item{multichain}{Boolean controlling whether to compute joint multi-chain distance matrices for observed chain combinations. Default is FALSE.}

\item{rdsOutputPath}{Path to the output directory for the RDS files containing the distance matrices. Default is "./tcrdist3DistanceMatrices/".}

\item{pythonExecutable}{Path to the python executable. Default is reticulate::py_exe().}

\item{debugTcrdist3}{Boolean controlling whether to run tcrdist3 in debug mode. Default is TRUE.}

\item{verbose}{Boolean controlling whether to display processing steps. Default is FALSE.}
}
\description{
This function runs the tcrdist3 pipeline on a set of CDR3 sequences in a Seurat Object.
}
\examples{
\dontrun{
  seuratObj_TCR<- RunTcrdist3(inputData = seuratObj@meta.data,
              chains = c("TRA", "TRB"),
              summarizeClones = T,
              imputeCloneNames = T,
              minimumClonesPerSubject = 2,
              multichain = FALSE,
              rdsOutputPath = "./tcrdist3DistanceMatrices/",
              pythonExecutable = reticulate::py_exe(),
              verbose = FALSE)

  # Example with multichain analysis
  seuratObj_TCR<- RunTcrdist3(inputData = seuratObj,
              chains = c("TRA", "TRB", "TRG", "TRD"),
              multichain = TRUE,
              verbose = TRUE)

    spikeInDataframe <- data.frame(CloneNames = rep(1:3),
                                 TRA_V = c("TRAV1-2", "TRAV1-2", "TRAV1-2"),
                                 TRA_J = c("TRAJ33", "TRAJ20", "TRAJ33"),
                                 TRA = c("CAVRDSNYQLIW", "CAVSLQDYKLSF", "CAVRDSNYQLIW"),
                                 TRB_V = c("TRBV6-4", "TRBV6-4", "TRBV6-4"),
                                 TRB_J = c("TRBJ1-1", "TRBJ2-1", "TRBJ2-3"),
                                 TRB = c("CASSAAAAAAAAFF", "CASSVVVVVVVVQF", "CASSWWWWWWWWQY"))
}

}
