% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrDist3Wrapper.R
\name{RunTcrdist3}
\alias{RunTcrdist3}
\title{RunTcrdist3}
\usage{
RunTcrdist3(
  seuratObj = NULL,
  metadata = NULL,
  organism = "human",
  formatMetadata = T,
  postFormattingMetadataCsvPath = "./tcrDist3Input.csv",
  chains = c("TRA", "TRB"),
  cleanMetadata = T,
  spikeInDataframe = NULL,
  summarizeClones = T,
  imputeCloneNames = T,
  minimumClonesPerSubject = 2,
  rdsOutputPath = "./tcrdist3DistanceMatrices/",
  pythonExecutable = reticulate::py_exe(),
  debugTcrdist3 = "True"
)
}
\arguments{
\item{seuratObj}{Seurat Object containing TCR information. If NULL, metadata must be provided.}

\item{metadata}{Data frame containing metadata. If NULL, seuratObj must be provided.}

\item{organism}{Organism to use for tcrdist3. Default is 'human'.}

\item{formatMetadata}{Boolean controlling whether to format the metadata for tcrdist3 using the internal FormatMetadataForTcrDist3 function. Default is TRUE.}

\item{postFormattingMetadataCsvPath}{Path to the output CSV file from FormatMetadataForTcrDist3. Default is './tcrDist3Input.csv'.}

\item{chains}{Vector of TCR chains to include in the analysis. Default is c("TRA", "TRB").}

\item{cleanMetadata}{Pass-through boolean controlling whether to clean the metadata by removing rows with NA values or commas in the specified chains. Default is TRUE.}

\item{summarizeClones}{Pass-through boolean controlling whether to summarize clones into a frequency (by SubjectId, TRA, TRB, TRA_V, and TRB_J). Default is TRUE.}

\item{imputeCloneNames}{Pass-through boolean controlling whether to impute clone names if they are missing. Existing clone names will be inherited. Default is TRUE.}

\item{minimumClonesPerSubject}{Minimum number of clones per subject to include in the analysis. Default is 2.}

\item{rdsOutputPath}{Path to the output directory for the RDS files containing the distance matrices. Default is "./tcrdist3DistanceMatrices/".}

\item{pythonExecutable}{Path to the python executable. Default is reticulate::py_exe().}

\item{debugTcrdist3}{String (to be passed to python and converted to boolean) controlling whether to run tcrdist3 in debug mode. Default is "True".}
}
\description{
This function runs the tcrdist3 pipeline on a set of CDR3 sequences in a Seurat Object.
}
\examples{
\dontrun{
  RunTcrdist3(seuratObj = seuratObj,
              metadata = NULL,
              formatMetadata = T,
              postFormattingMetadataCsvPath = './tcrDist3Input.csv',
              chainsString = "alpha,beta",
              cleanMetadata = T,
              summarizeClones = T,
              imputeCloneNames = T,
              minimumClonesPerSubject = 2,
              rdsOutputPath = "./tcrdist3DistanceMatrices/",
              pythonExecutable = reticulate::py_exe(),
              debugTcrdist3 = "True")
}
}
